+++ android-lifecycle-runtime/src/main/java/dispatch/android/lifecycle/LifecycleCoroutineScope.kt
@@ -13,10 +13,10 @@
  * limitations under the License.
  */
 
-package dispatch.android
+package dispatch.android.lifecycle
 
 import androidx.lifecycle.*
-import dispatch.android.internal.*
+import dispatch.android.lifecycle.internal.*
 import dispatch.core.*
 import kotlinx.coroutines.*
